requirements.txt
python-telegram-bot==20.7
feedparser==6.0.12
apscheduler==3.11.2
